Add files via upload
As the developer of this code, I was required to use the Android Emulator feature of Android Studio to test my code. My app should work for users that both do and do not have accounts. This means you should focus on the following functionality:
-The app should check the username and password against the database when the user attempts to log in.
-If the user has never logged into the application before, the user should be able to create a new login and password. The application needs to save these to a table in the database.
The main feature required to develop my code was Visual Studio and  I had to explore different tools like IDEs to help me achieve my goals. My focus stayed majorly on developing a user friendly app by making sure I tested  my code often with the Android Emulator. My completed database was to allow a user to accomplish the following:
: The user should be able to add items to a database.
: The user should be able to remove items from a database.
: The user should be able to change the value associated with individual database items (e.g. the number of a specific item in an inventory or the date of an event).
: The user should be able to view all of the database items displayed as a grid.
To be productive and innovative, I allowed the use of IntelliSense code suggestions which gave entire lines of code with suggestions to help me make quick improvements to my code using light bulbs. My biggest skill that I'd say successfully showcased is the design part , which allowed me to be as unique as possible to meet user requirements.